split drawing sprite frames and split saving into two seperate files
